auto-test/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
assertpy==1.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.8.0
execnet==2.1.1
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
h11==0.16.0
httplib2==0.31.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.14.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==7.4.3
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-reportlog==0.3.0
pytest-xdist==3.5.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
timeout-decorator==0.5.0
trio==0.30.0
trio-websocket==0.12.2
uritemplate==4.2.0
urllib3==2.5.0
webdriver-manager==4.0.1
wsproto==1.2.0
